Fix #93: Crypto 3.0: Unify error handling in REST endpoints Introduce request/response objects for /pa/v3/activation/status and /pa/v3/activation/status Move service status endpoint to v3
