+++ powerauth-restful-model/src/main/java/io/getlime/security/powerauth/rest/api/model/request/v3/ActivationStatusRequest.java
@@ -1,0 +1,48 @@
+/*
+ * PowerAuth integration libraries for RESTful API applications, examples and
+ * related software components
+ *
+ * Copyright (C) 2017 Lime - HighTech Solutions s.r.o.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package io.getlime.security.powerauth.rest.api.model.request.v3;
+
+/**
+ * Request object for /pa/v3/activation/status end-point
+ *
+ * @author Petr Dvorak, petr@wultra.com
+ *
+ */
+public class ActivationStatusRequest {
+
+    private String activationId;
+
+    /**
+     * Get activation ID
+     * @return Activation ID
+     */
+    public String getActivationId() {
+        return activationId;
+    }
+
+    /**
+     * Set activation ID
+     * @param activationId Activation ID
+     */
+    public void setActivationId(String activationId) {
+        this.activationId = activationId;
+    }
+
+}
